backend/requirements.txt
cryptography==46.0.3
Flask==3.1.2
flask_cors==6.0.1
fpdf==1.7.2
pandas==2.3.3
Pillow==12.0.0
psycopg2_binary==2.9.11
bcrypt==4.1.2
python_magic==0.4.27
Requests==2.32.5
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv
flask-marshmallow
openpyxl==3.1.3